Use time.time instead of datetime.utcnow
According to the document of time.time(), "Return the current time in
seconds since the Epoch" and "Seconds Since the Epoch" does not depend
on timezone [1].  So it's better to have less code to store the same
thing.  Also, IPython's wall time of %time is also calculated by
time.time().

[1] http://stackoverflow.com/a/20035913