Reverse call back order in Defer.call()
This is how Go's defer works.  In the last commit, I could have called
_wait_bg via defer if callbacks are called in this way.  It means that
this allows user to implement save_maybe using defer.  As I think
save_maybe is more robust in the way it is implemented in the last
commit, I'll keep it as-is.  But this is the motivation for changing the
Defer.call() order to FILO.